wrote levy flight but didn't test it yet tho everything is set up nicely so all futher tinkerigs will be a breeze
acute problems b4 prez:
- plots plots and more plots
- creating many different environments and stories around them
- actual prez

there remain 2 non trivial coding tasks:
1)
write the ann/whatever
thing to swith states and tune internal parameters (things like force
constants, ni and so on) based on the environment sensors (counting
trees, items, whatever in a limited time frame, use that as inputs to
ann).  most likely you'll be able to get away with the simplest possible
thing with like 3 neurons in a single layer, but check.
furthermore check wether that's the best computing arh
---> after prez.

2) since you'll need to run the simulation a few times to get the evaluation
for a single ann setting you'll need to write a multiprocessing program
to run the multiple sims. preferably make it networked as well to
harness multiple computers. you'll possibly need to go a bit deeper
than described in terms of task decomposition,
but the details will be more evident when you actually start coding the
thing.
(btw this is a must for 1) as a single simulation will take 5~10mins,
depending on what sim params we choose)
---> (obv) after prez